requirements.txt
amqp==5.1.1
annotated-types==0.7.0
anthropic==0.45.0
anyio==4.3.0
--extra-index-url=https://pypi.apryse.com
apryse-sdk
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.1
Automat==22.10.0
billiard==4.1.0
boto3==1.28.9
botocore==1.31.9
Brotli==1.1.0
celery==5.3.4
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==15.1.0
contourpy==1.2.0
cron_descriptor==1.4.0
cryptography==41.0.1
cssselect2==0.7.0
cycler==0.12.1
daphne==4.0.0
Deprecated==1.2.14
distro==1.9.0
Django==4.1.7
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==23.1
django-rest-passwordreset==1.3.0
django-silk==5.1.0
django-storages==1.13.2
django-timezone-field==6.0.1
djangorestframework==3.14.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
exceptiongroup==1.2.1
fonttools==4.43.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jmespath==1.0.1
kiwisolver==1.4.5
kombu==5.3.2
lxml==4.9.3
matplotlib==3.8.1
mysqlclient==2.1.1
neverbounce_sdk==4.3.0
numpy==1.26.0
openai==1.30.2
openpyxl==3.1.2
oscrypto==1.3.0
packaging==23.2
pandas==2.1.1
pdfkit==1.0.0
pdfminer==20191125
Pillow==9.4.0
prompt-toolkit==3.0.39
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pydyf==0.8.0
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
PyJWT==2.6.0
PyMuPDF==1.24.7
pyOpenSSL==23.2.0
pyparsing==3.1.1
pypdf==3.16.4
PyPDF2==3.0.1
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.7.1
PyYAML==6.0.1
qrcode==7.4.2
redis==4.3.4
reportlab==3.6.13
requests==2.30.0
s3transfer==0.6.1
service-identity==21.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.3
svglib==1.5.1
tinycss2==1.2.1
tqdm==4.66.4
Twisted==22.10.0
twilio==9.3.0
txaio==23.1.1
typing_extensions==4.12.0
tzdata==2022.7
tzlocal==5.2
uritools==4.0.2
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.8
webencodings==0.5.1
wrapt==1.15.0
xhtml2pdf==0.2.11
xlsxwriter==3.2.0
zope.interface==6.0
zopfli==0.2.3
num2words==0.5.13